Create tidalSeq
tidalSequencer

this code is for an arduino (written for nano) interfacing with a modular synth, intended primarily as a drum sequencer. A schematic for the protection circuit is available. The sequencer reads 2 pots (with additional CV input for tide pot) and looks at a preset pattern of hits, documented as 0-9 probability of playing. The tide value represents a water level, with the pattern like a row of posts of different height in the water - as the tide comes up, only the taller posts still show (play), while a low tide (pot turned clockwise) sees every beat play. The 2nd pot controls the swell level, which adds an amount of randomness to the probability of that drum playing - turning this up makes it glitchy and random. The pattern sets the bar length. 
There is also a Euclidean mode which uses the 2 pots differently - the swell pot controls the bar length, and the tide pot controls the proportion of hits for that bar length, from 0 to all.  This distributes a number of hits as evenly as possible a bar of variable length, giving a variety of world rhythmic figures. Modes are set to tidal or Euclidean per channel. 
5 buttons give various functions, allowing shuffle, pattern scrolling and randomising, plus storing & recall of good combinations and an undo facility. LEDs signal which channel is being edited.