Create solenoidDecay
This is an arduino function intended to run on an ATTiny25/45/85. 
2 pots control freq & decay
gate in -> optocoupler & triggers signal in MCU
MCU outputs to TIP120 Darlington to give power to solenoid
single gate triggers multiple hits of specified duration & frequency, endlessly, or for a specified decayTime. 
(? add PWM decaying over time)

This triggers single or multiple hits on a cymbal, for instance, from 1 trigger - listen back via a contact mic?